+++ src/components/consts.js
@@ -1,0 +1,13 @@
+const FILM_COUNT = 5;
+const FILM_EXTRA_COUNT = 2;
+
+const Place = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+const ExtraFilmTitle = {
+  TOP_RATED: `Top rated`,
+  TOP_COMMENTED: `Most commented`
+};
+
+export {FILM_COUNT, FILM_EXTRA_COUNT, Place, ExtraFilmTitle};
