+++ src/components/show-more-btn.js
@@ -1,1 +1,1 @@
-export const createShowMoreBtn = () => (`<button class="films-list__show-more">Show more</button>`);
+export const createShowMoreBtn = () => `<button class="films-list__show-more">Show more</button>`;
