+++ src/components/statistics.js
@@ -1,1 +1,1 @@
-export const createStatistics = () => (`<p>130 291 movies inside</p>`);
+export const createStatistics = () => `<p>130 291 movies inside</p>`;
