+++ src/main.js
@@ -1,37 +1,18 @@
-import {createProfile} from "./components/profile.js";
-import {createNavigation} from "./components/navigation.js";
-import {createFilter} from "./components/filter.js";
-import {createFilmCard} from "./components/film-card.js";
-import {createFilmDetails} from "./components/film-details.js";
-import {createFilmsList} from "./components/films-list.js";
-import {createFilmsListExtra} from "./components/films-list-extra.js";
-import {createShowMoreBtn} from "./components/show-more-btn.js";
-import {createStatistics} from "./components/statistics.js";
-
-const FILM_COUNT = 5;
-const FILM_EXTRA_COUNT = 2;
-
-const Place = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-const ExtraFilmTitle = {
-  TOP_RATED: `Top rated`,
-  TOP_COMMENTED: `Most commented`
-};
+import {createProfile} from "./components/profile";
+import {createNavigation} from "./components/navigation";
+import {createFilter} from "./components/filter";
+import {createFilmDetails} from "./components/film-details";
+import {createFilmsList} from "./components/films-list";
+import {createFilmsListExtra} from "./components/films-list-extra";
+import {createShowMoreBtn} from "./components/show-more-btn";
+import {createStatistics} from "./components/statistics";
+import {render, renderFilmCards} from "./components/utils";
+import {Place, ExtraFilmTitle, FILM_COUNT, FILM_EXTRA_COUNT} from "./components/consts";
 
 const siteHeader = document.querySelector(`.header`);
 const siteMain = document.querySelector(`.main`);
 const siteFooter = document.querySelector(`.footer`);
 const siteFooterStat = siteFooter.querySelector(`.footer__statistics`);
-
-const render = (container, template, place) => container.insertAdjacentHTML(place, template);
-
-const renderFilmCards = (container, count) => {
-  for (let i = 0; i < count; i++) {
-    render(container, createFilmCard(), Place.BEFOREEND);
-  }
-};
 
 const init = () => {
   render(siteHeader, createProfile(), Place.BEFOREEND);
